[KYV-1064] do rollback

+++ core/src/main/java/org/helsinki/paytrail/model/payments/PaymentItem.java
@@ -10,7 +10,8 @@ public class PaymentItem {
     private int unitPrice;
     private int units;
     // for decimal handling (KYV-1064)
-    private double vatPercentage;
+//    private double vatPercentage;
+    private int vatPercentage;
     private String productCode;
     private String description;
     private String orderId;
